feat: web interface
